+++ service/azureconfig/v1/cloudconfig/types.go
@@ -2,7 +2,9 @@ package cloudconfig
 
 import (
 	providerv1alpha1 "github.com/giantswarm/apiextensions/pkg/apis/provider/v1alpha1"
+
 	"github.com/giantswarm/azure-operator/client"
+	"github.com/giantswarm/azure-operator/service/azureconfig/config"
 	"github.com/giantswarm/azure-operator/service/azureconfig/v1/key"
 )
 
@@ -30,7 +32,7 @@ type cloudProviderConfFileParams struct {
 	VnetName          string
 }
 
-func newCloudProviderConfFileParams(azure key.Azure, azureConfig client.AzureConfig, obj providerv1alpha1.AzureConfig) cloudProviderConfFileParams {
+func newCloudProviderConfFileParams(azure config.Azure, azureConfig client.AzureConfig, obj providerv1alpha1.AzureConfig) cloudProviderConfFileParams {
 	return cloudProviderConfFileParams{
 		AzureCloudType:    key.AzureCloudType(obj),
 		Location:          azure.Location,
