+++ service/azureconfig/v1/resource/deployment/resource.go
@@ -14,6 +14,7 @@ import (
 	"github.com/giantswarm/operatorkit/framework"
 
 	"github.com/giantswarm/azure-operator/client"
+	"github.com/giantswarm/azure-operator/service/azureconfig/config"
 	"github.com/giantswarm/azure-operator/service/azureconfig/v1/cloudconfig"
 	"github.com/giantswarm/azure-operator/service/azureconfig/v1/key"
 )
@@ -30,7 +31,7 @@ type Config struct {
 	CloudConfig   *cloudconfig.CloudConfig
 	Logger        micrologger.Logger
 
-	Azure       key.Azure
+	Azure       config.Azure
 	AzureConfig client.AzureConfig
 	// TemplateVersion is the ARM template version. Currently is the name
 	// of the git branch in which the version is stored.
@@ -42,7 +43,7 @@ type Resource struct {
 	cloudConfig   *cloudconfig.CloudConfig
 	logger        micrologger.Logger
 
-	azure           key.Azure
+	azure           config.Azure
 	azureConfig     client.AzureConfig
 	templateVersion string
 }
