+++ service/azureconfig/v1/resource/deployment/resource_test.go
@@ -11,12 +11,12 @@ import (
 	"github.com/giantswarm/randomkeys/randomkeystest"
 
 	"github.com/giantswarm/azure-operator/client/fakeclient"
+	"github.com/giantswarm/azure-operator/service/azureconfig/config"
 	"github.com/giantswarm/azure-operator/service/azureconfig/v1/cloudconfig"
-	"github.com/giantswarm/azure-operator/service/azureconfig/v1/key"
 )
 
-var testAzure = key.Azure{
-	HostCluster: key.AzureHostCluster{
+var testAzure = config.Azure{
+	HostCluster: config.AzureHostCluster{
 		CIDR:           "10.0.0.0/8",
 		ResourceGroup:  "test-group",
 		VirtualNetwork: "test-vnet",
